Add setiing FIFO threshold

+++ src/halk/halk_spi.h
@@ -2,4 +2,4 @@
 
 #include <stm32h7xx_ll_spi.h>
 
-void halk_spi_transmit(SPI_TypeDef *spi, uint8_t *data, uint32_t size);
+void halk_spi_transmit(SPI_TypeDef* spi, uint8_t* data, uint16_t size);
